refactor : batch의 database 관련 폴더 분리
